BACKLOG-12343 : move Picker related logic in the same component using it
