+++ src/javascript/JContent/ContentRoute/ContentLayout/ContentTrees/ContentTree/ContentTree.jsx
@@ -4,6 +4,8 @@ import {Picker} from '@jahia/react-apollo';
 import {PredefinedFragments} from '@jahia/apollo-dx';
 import {TreeView} from '@jahia/moonstone';
 import gql from 'graphql-tag';
+import JContentConstants from '../../../../JContent.constants';
+import {Collections, File, FolderSpecial, Setting} from '@jahia/moonstone/dist/icons';
 
 const PickerItemsFragment = {
     mixinTypes: {
@@ -38,6 +40,72 @@ const PickerItemsFragment = {
     }
 };
 
+function getIcon(mode) {
+    switch (mode) {
+        case JContentConstants.mode.PAGES:
+            return <File size="small"/>;
+        case JContentConstants.mode.CONTENT_FOLDERS:
+            return <FolderSpecial size="small"/>;
+        case JContentConstants.mode.MEDIA:
+            return <Collections size="small"/>;
+        case JContentConstants.mode.APPS:
+            return <Setting size="small"/>;
+        default:
+            return <File size="small"/>;
+    }
+}
+
+function getParentPath(path) {
+    return path.substr(0, path.lastIndexOf('/'));
+}
+
+function findInTree(tree, id) {
+    for (var i = 0; i < tree.length; i++) {
+        if (tree[i].id === id) {
+            return tree[i];
+        }
+
+        let result = findInTree(tree[i].children, id);
+        if (result) {
+            return result;
+        }
+    }
+}
+
+function convertPathsToTree(pickerEntries, mode) {
+    let tree = [];
+    if (pickerEntries.length === 0) {
+        return tree;
+    }
+
+    let rootElement = {
+        id: pickerEntries[0].path,
+        label: pickerEntries[0].node.displayName,
+        hasChildren: pickerEntries[0].hasChildren,
+        parent: getParentPath(pickerEntries[0].path),
+        iconStart: getIcon(mode),
+        children: []
+    };
+    tree.push(rootElement);
+    for (let i = 1; i < pickerEntries.length; i++) {
+        let parentPath = getParentPath(pickerEntries[i].path);
+        let element = {
+            id: pickerEntries[i].path,
+            label: pickerEntries[i].node.displayName,
+            hasChildren: pickerEntries[i].hasChildren,
+            parent: parentPath,
+            iconStart: getIcon(mode),
+            children: []
+        };
+        let parent = findInTree(tree, parentPath);
+        if (parent !== undefined) {
+            parent.children.push(element);
+        }
+    }
+
+    return tree;
+}
+
 class ContentTree extends React.Component {
     constructor(props) {
         super(props);
@@ -45,7 +113,7 @@ class ContentTree extends React.Component {
     }
 
     render() {
-        let {rootPath, openPaths, path, handleOpen, handleSelect, lang, openableTypes, selectableTypes, setRefetch, mode, convertPathsToTree} = this.props;
+        let {rootPath, openPaths, path, handleOpen, handleSelect, lang, openableTypes, selectableTypes, setRefetch, mode} = this.props;
         return (
             <Picker
                 ref={this.picker}
@@ -85,7 +153,6 @@ ContentTree.propTypes = {
     mode: PropTypes.string.isRequired,
     openPaths: PropTypes.arrayOf(PropTypes.string).isRequired,
     openableTypes: PropTypes.array.isRequired,
-    convertPathsToTree: PropTypes.func.isRequired,
     path: PropTypes.string.isRequired,
     rootPath: PropTypes.string.isRequired,
     selectableTypes: PropTypes.array.isRequired,
